+++ examples/simple.rs
@@ -79,6 +79,9 @@ fn main() {
     base_widget.set_origin(50, 80);
     base_widget.set_size(200, 200);
     base_widget.set_color([0.5, 0.5, 0.5, 1.0]);
+    base_widget.on_mouse_entered(Box::new(|widget_id| {
+        eprintln!("Mouse entered widget {}", widget_id);
+    }));
     prod.widget_store.add_widget(Box::new(base_widget));
 
     let mut box_widget = BoxWidget::new();